Refactored traineeController
